# docs/backend/backend_python/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PictoPy",
    "description": "The API calls to PictoPy are done via HTTP requests. This backend is built using FastAPI.",
    "version": "0.1.0",
    "contact": {
      "name": "PictoPy Postman Collection",
      "url": "https://www.postman.com/aossie-pictopy/pictopy/overview"
    }
  },
  "servers": [
    {
      "url": "http://localhost:52123",
      "description": "Local Development server"
    }
  ],
  "paths": {
...
    "/avatars/upload": {
      "post": {
        "tags": [
          "Avatar"
        ],
        "summary": "Upload Avatar",
        "description": "Upload and set user avatar",
        "operationId": "upload_avatar_avatars_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_avatars_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/avatars/uploads/{filename}": {
      "get": {
        "tags": [
          "Avatar"
        ],
        "summary": "Get Avatar",
        "description": "Serve uploaded avatar files",
        "operationId": "get_avatar_avatars_uploads__filename__get",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_upload_avatar_avatars_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_avatar_avatars_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}